decisions were made
